Add files via upload
In this project, I created a website for the Top 3 movies of all times, providing their titles, directors' names, and summaries. When the user click on the Title of the movie, a new tab will pop open and display the title, director's name, summary and the poster of the movie. The website is in Turkish.